3 dataset architecture, untested model abstraction

+++ processing/corruption-checker.py
@@ -4,7 +4,7 @@ import glob, os
 
 broken = []
 
-for fname in glob.glob('data/proc/new/224/**/*.png') + glob.glob('data/proc/old/224/**/*.png') + glob.glob('data/proc/aug/224/**/*.png'):
+for fname in glob.glob('data/proc/new/224/**/*.png') + glob.glob('data/proc/old/224/**/*.png'):
 	try:
 		Image.open(fname).verify()
 	except:
@@ -14,6 +14,7 @@ for fname in glob.glob('data/proc/new/224/**/*.png') + glob.glob('data/proc/old/
 if input('delete all? y/n: ') == 'y':
 	for fname in broken:
 		os.remove(fname)
-		os.remove(fname.replace('224','299'))
+		if os.file.exists(fname.replace('224','299')):
+			os.remove(fname.replace('224','299'))
 else:
 	print(broken)
